Minor UI changes
- Removed usage of Lato font, as it has some problems in some setups.
- For the moment, the Lato font will only used in iOS.
- However, I will search for another font and upload it soon.

+++ src/device_manager.h
@@ -26,7 +26,7 @@ class DeviceManager : public QObject {
 
 public:
     Q_INVOKABLE bool isMobile();
-    Q_INVOKABLE int ratio(int value);
+    Q_INVOKABLE qreal ratio(int value);
 };
 
 #endif
